EmailContentProvider: An agent for polling for new email, and serving requests for email.
This new agent is not yet exercised and is effectively dead code for
now. Wiring up will be in another patch.

Modules can connect to this agent which provides the EmailContentProvider
service. This service (email_content_provider.fidl) serves requests for
a user's email, and provides a way to be notified of new email.

The EmailContentProvider service periodically (every 1 minute) fetches
email history. If changes exist, all of the email is fetched again, and
this is now the email that is served out. A client of
EmailServiceProvider will need to fetch threads() again.
There is a lot to improve here (per-label notifications)

Change-Id: I63ca1e7037e6f5ad87b33eae01fda804ec428984